# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "API for chess tournament administration system.",
        "contact": {
            "email": "xdudasko@stuba.sk, xdemeterp@stuba.sk"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://mtaa_chess_api.test/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get list of players",
                "description": "Returns list of players",
                "operationId": "getPlayersList",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Jozef Nový",
                                        "elo": 2650,
                                        "title": "GM",
                                        "category": "CH24"
                                    },
                                    {
                                        "id": 2,
                                        "name": "František Starý",
                                        "elo": 2600,
                                        "title": "GM",
                                        "category": "CH24"
                                    },
                                    {
                                        "id": 3,
                                        "name": "Samuel Halčin",
                                        "elo": 2477,
                                        "title": "IM",
                                        "category": "CH18"
                                    },
                                    {
                                        "id": 3,
                                        "name": "Anna Lizáková",
                                        "elo": 2311,
                                        "title": "FM",
                                        "category": "D18"
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Store new player",
                "description": "Returns player data",
                "operationId": "storePlayer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "František Nový"
                                    },
                                    "elo": {
                                        "type": "integer",
                                        "example": "1000"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "CH20"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "null"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "František Nový"
                                        },
                                        "elo": {
                                            "type": "integer",
                                            "example": "1000"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "CH20"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "created_at": {
                                            "type": "timestamp",
                                            "example": "2021-05-06 09:00:00"
                                        },
                                        "updated_at": {
                                            "type": "timestamp",
                                            "example": "2021-05-06 09:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/players/storeBulk": {
            "put": {
                "tags": [
                    "Players"
                ],
                "summary": "Store list of players",
                "description": "Returns whether operation was successful.",
                "operationId": "storeBulkPlayers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": [
                                {
                                    "name": "Jozef Nový",
                                    "elo": 2650,
                                    "title": "GM",
                                    "category": "CH24"
                                },
                                {
                                    "name": "František Starý",
                                    "elo": 2600,
                                    "title": "GM",
                                    "category": "CH24"
                                },
                                {
                                    "name": "Samuel Halčin",
                                    "elo": 2477,
                                    "title": "IM",
                                    "category": "CH18"
                                },
                                {
                                    "name": "Anna Lizáková",
                                    "elo": 2311,
                                    "title": "FM",
                                    "category": "D18"
                                }
                            ]
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Something went wrong."
                    }
                }
            }
        },
        "/standings": {
            "get": {
                "tags": [
                    "Standings"
                ],
                "summary": "Get standings",
                "description": "Returns standings after last round if not specified otherwise.",
                "operationId": "getStandings",
                "parameters": [
                    {
                        "name": "group_by_category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "bool"
                        }
                    },
                    {
                        "name": "round",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Jozef Nový",
                                        "elo": 2650,
                                        "title": "GM",
                                        "category": "CH24",
                                        "points": 3
                                    },
                                    {
                                        "id": 2,
                                        "name": "František Starý",
                                        "elo": 2600,
                                        "title": "GM",
                                        "category": "CH24",
                                        "points": 2
                                    },
                                    {
                                        "id": 3,
                                        "name": "Samuel Halčin",
                                        "elo": 2477,
                                        "title": "IM",
                                        "category": "CH18",
                                        "points": 5.5
                                    },
                                    {
                                        "id": 3,
                                        "name": "Anna Lizáková",
                                        "elo": 2311,
                                        "title": "FM",
                                        "category": "D18",
                                        "points": 3.5
                                    }
                                ]
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/users/checkPlayerId": {
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Check if the typed id is a valid player id.",
                "description": "Returns true if the player can attend the tournament.",
                "operationId": "checkPlayerId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "query",
                        "description": "Tournament hash",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns true if the player can attend the tournament.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "bool",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}